MSYS2: adapt waf script: skip gtk2/gtk3, translate wxConfigDir
I needed to translate the wxConfigDir because otherwise it would not
be recognized by waf executed by MINGW python.  I am not sure if this
is the best solution, maybe using waf from MSYS python instead of
MINGW python would be better, but MSYS2 does not have a waf package
for MSYS python.